+++ src/basins/basins_utilities.jl
@@ -19,14 +19,14 @@ end
 # Sampled Basin of Attraction with sampled points as a vector of vectors
 function SampledBasinsOfAttraction(
         points_ids::Vector{ID}, attractors::Dict{AK, S}, sampled_points::Vector{U};
-        tree = KDTree, metric = Euclidean(), ss_kwargs...
+        ss_kwargs...
     ) where {ID, D, T, V <: AbstractVector, U <: AbstractVector, AK, S <: StateSpaceSet{D, T, V}}
     sampled_set = StateSpaceSet(sampled_points)
     eltype(S) != V && error(
         "The attractor points and sampled points must be represented " *
             "by the same type of vector, that is they must have the same element type and length"
     )
-    return SampledBasinsOfAttraction(points_ids, attractors, sampled_set; tree = tree, metric = metric, ss_kwargs = ss_kwargs)
+    return SampledBasinsOfAttraction(points_ids, attractors, sampled_set; ss_kwargs...)
 end
 
 #########################################################################################
@@ -37,8 +37,8 @@ end
 
 Compute the full basins of attraction as identified by the given `mapper`,
 which includes a reference to a [`DynamicalSystem`](@ref) and return them
-along with (perhaps approximated) found attractors contained within the 
-[`ArrayBasinsOfAttraction`](@ref) structure. 
+along with (perhaps approximated) found attractors contained within the
+[`ArrayBasinsOfAttraction`](@ref) structure.
 
 `grid` is a tuple of ranges defining the grid of initial conditions that partition
 the state space into boxes with size the step size of each range.
@@ -54,7 +54,7 @@ the partitioning happens directly on the hyperplane the Poincaré map operates o
 `labels` returned by [`basins_fractions`](@ref) and simply assigns them to a full array
 corresponding to the state space partitioning indicated by `grid`.
 
-Note that, to ensure backwards compatibility the return type can be decomposed such that 
+Note that, to ensure backwards compatibility the return type can be decomposed such that
 `basins, attractors = boa`.
 
 See also [`convergence_and_basins_of_attraction`](@ref).
@@ -73,15 +73,15 @@ end
 
 Compute the full basins of attraction as identified by the given `mapper`,
 which includes a reference to a [`DynamicalSystem`](@ref) and return them
-along with (perhaps approximated) found attractors contained within a 
+along with (perhaps approximated) found attractors contained within a
 [`SampledBasinsOfAttraction`](@ref) object.
 
 The initial conditions `ics`, and the keyword arguments `kwargs` are the same
 as in [`basins_fractions`](@ref) with the same function signature. This function
-is a small convenience wrapper which uses the sampled initial conditions and their 
+is a small convenience wrapper which uses the sampled initial conditions and their
 corresponding labels, from `basins_fractions`, to construct a [`SampledBasinsOfAttraction`](@ref)
 
-Note that, as with the other `basins_of_attraction` function, the return can be decomposed: 
+Note that, as with the other `basins_of_attraction` function, the return can be decomposed:
 `basins, attractors = boa`.
 
 """
@@ -126,7 +126,7 @@ Return a dictionary that maps attractor IDs to their relative fractions.
 Optionally you may give a vector of `ids` to calculate the fractions of only
 the chosen ids (by default `ids = unique(basins)`).
 
-The second function signature exists for backwards compatibility. 
+The second function signature exists for backwards compatibility.
 
 In [Menck2013](@cite) the authors use these fractions to quantify the stability of a basin of
 attraction, and specifically how it changes when a parameter is changed.
@@ -231,9 +231,9 @@ end
 
 Given a `point`, `map_to_basin` interpolates to which basin it should belong.
 
-For [`ArrayBasinsOfAttraction`](@ref) this finds the label of the closest grid cell. 
+For [`ArrayBasinsOfAttraction`](@ref) this finds the label of the closest grid cell.
 
-For [`SampledBasinsOfAttraction`](@ref) this finds the label corresponding to the 
+For [`SampledBasinsOfAttraction`](@ref) this finds the label corresponding to the
 nearest neighbor using `Neighborhood.jl`
 
 For developing a new `BasinOfAttraction` subtype extend the internal function `map_to_domain`,